🔨 chore: move the `useSwitchSession` file to `hooks`
